feat: real pipeline cancellation via PipelineHandle + Arc<AtomicBool>
- ChainRunner gains cancelled: Arc<AtomicBool>, checked in recv loop every 100ms
- Pipeline::compile() accepts cancellation token, passes to all ChainRunners
- Pipeline::run() uses thread::spawn, returns PipelineHandle with .cancel() and .join()
- merge_inputs uses thread::spawn instead of thread::scope
- PipelineEvent::Cancelled added
- Dispatcher stores background_tasks: HashMap<TabId, PipelineHandle>
- Dispatcher::cancel_background(tab_id) cancels and removes handle
- dispatch_blur_preview cancels previous background pipeline before dispatching new
- CloseTab calls dispatcher.cleanup_tab() which cancels any running pipeline
- CacheWriter uses dynamic bytes_per_pixel() instead of hardcoded 4

+++ pixors-executor/src/sink/cache_writer.rs
@@ -67,8 +67,8 @@ impl Consumer for CacheWriterConsumer {
                 return Err(Error::internal("CacheWriter requires CPU tiles"));
             }
         };
-        let bpp = tile.meta.format.bytes_per_pixel() as usize;
-        let expected = (w as usize * h as usize * bpp) as usize;
+        let bpp = tile.meta.format.bytes_per_pixel();
+        let expected = w as usize * h as usize * bpp;
         if data.len() != expected {
             return Err(Error::internal(format!(
                 "CacheWriter tile size mismatch: expected {expected} bytes, got {}",
